tests: move direnv env parsing in to its own module file
