+++ tests/integration/bug23_gopath.rs
@@ -1,4 +1,5 @@
-use direnvtestcase::{DirenvTestCase, DirenvValue};
+use direnv::DirenvValue;
+use direnvtestcase::DirenvTestCase;
 use std::env;
 
 #[test]
