+++ tests/integration/direnvtestcase.rs
@@ -1,13 +1,13 @@
 //! Implement a wrapper around setup and tear-down of Direnv-based test
 //! cases.
 
+use direnv::DirenvEnv;
 use lorri::{
     build_loop::{BuildError, BuildLoop, BuildResults},
     ops::direnv,
     project::Project,
     roots::Roots,
 };
-use std::collections::HashMap;
 use std::fs::File;
 use std::io::Write;
 use std::path::PathBuf;
@@ -88,39 +88,4 @@ impl DirenvTestCase {
 
         d
     }
-}
-
-/// The resulting environment Direnv after running Direnv. Note:
-/// Direnv returns `{ "varname": null, "varname": "something" }`
-/// so the value type is `Option<String>`. This makes `.get()`
-/// operations clunky, so be prepared to check for `Some(None)` and
-/// `Some(Some("val"))`.
-#[derive(Deserialize, Debug)]
-pub struct DirenvEnv(HashMap<String, Option<String>>);
-
-impl DirenvEnv {
-    /// Get an environment value with a borrowed str in the deepest Option.
-    /// Makes asserts nicer, like:
-    ///
-    ///    assert!(env.get_env("foo"), Value("bar"));
-    pub fn get_env<'a, 'b>(&'a self, key: &'b str) -> DirenvValue {
-        match self.0.get(key) {
-            Some(Some(val)) => DirenvValue::Value(&val),
-            Some(None) => DirenvValue::Unset,
-            None => DirenvValue::NotSet,
-        }
-    }
-}
-
-/// Environemnt Values from Direnv
-#[derive(Debug, PartialEq)]
-pub enum DirenvValue<'a> {
-    /// This variable will not be modified.
-    NotSet,
-
-    /// This variable will be unset when entering direnv.
-    Unset,
-
-    /// This variable will be set to exactly.
-    Value(&'a str),
 }
